feat: useSearchData 커스텀 훅 구현
